requirements.txt
fastapi==0.68.1
pydantic==1.8.2
uvicorn==0.15.0
uuid
dotenv
google-auth
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pandas
requests
numpy
torch
transformers
scikit-learn
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
prometheus-client==0.12.0
python-dotenv==0.19.0
redis==4.3.4